Use uint32_t for callbacks

+++ include/Instrumentation/Callbacks.h
@@ -10,13 +10,13 @@ namespace llvm {
 
 namespace mull {
 
-  extern "C" void mull_enterFunction(void **trampoline, uint64_t functionIndex);
-  extern "C" void mull_leaveFunction(void **trampoline, uint64_t functionIndex);
+  extern "C" void mull_enterFunction(void **trampoline, uint32_t functionIndex);
+  extern "C" void mull_leaveFunction(void **trampoline, uint32_t functionIndex);
 
   class Callbacks {
   public:
     void injectCallbacks(llvm::Function *function,
-                         uint64_t index,
+                         uint32_t index,
                          llvm::Value *infoPointer,
                          llvm::Value *offset);
