+++ include/Instrumentation/InstrumentationInfo.h
@@ -6,7 +6,7 @@
 namespace mull {
 struct InstrumentationInfo {
   InstrumentationInfo() :callTreeMapping(nullptr), callstack() {}
-  uint64_t *callTreeMapping;
-  std::stack<uint64_t> callstack;
+  uint32_t *callTreeMapping;
+  std::stack<uint32_t> callstack;
 };
 }
